Make the lexer context-sensitive.
The main reason is that keys are allowed to be anything,
including what would normally be reserved keywords: false, nan etc.
Even `2 = false` is a valid TOML document.

Since we are adding state to the lexer anyway,
it seems sensible to also make newline token generation context-sensitive:
in situations when newline is treated as whitespace,
the lexer will no longer emit it as a token,
which simplifies the grammar and reduces conflicts.