Integrate duplicate path checking into from_statements.
It's now based on the concept of "seen paths".
When a table header is seen and when a table array header
is seen for the first time, it's added to the list of seen paths.

Every new path is checked against that list.
New values are added to the head of the list, so it's constant time,
unlike the O(N) append of the old approach.